Use Unicode characters by default on Windows in src/print-tree.lisp
The motivation for the special case, that is #+win32 nil, was an
experiment with SBCL in some Windows environment. In that particular
environment, the Unicode characters did not work right away so using
the ASCII characters seemed like the safest choice. Now people report
that the problem no longer exists.

* src/print-tree.lisp (header): updated copyright
  (*use-unicode?*): removed special default for #+win32

fixes #1